requirements.txt
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
av==14.1.0
Babel==2.8.0
bcrypt==3.2.0
blessed==1.20.0
blessings==1.7
blinker==1.4
bpython==0.24
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
ctranslate2==4.5.0
curtsies==0.4.2
cwcwidth==0.1.9
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.20.2
huggingface-hub==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.7
MarkupSafe==2.0.1
matplotlib==3.10.0
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
numpy==2.2.1
oauthlib==3.2.0
onnxruntime==1.20.1
packaging==24.1
parso==0.8.1
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycurl==7.44.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.9.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
pyxdg==0.28
PyYAML==5.4.1
requests==2.32.3
SCons==4.0.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sqlparse==0.5.1
ssh-import-id==5.11
sympy==1.13.3
systemd-python==234
tokenizers==0.21.0
tqdm==4.67.1
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
watchdog==2.1.6
wcwidth==0.2.13
whitenoise==6.7.0
xdg==5
zipp==1.0.0
zope.interface==5.4.0